Fix bug resulting in incorrect rendering
A major bug was found and fixed that resulted in the wrong cell being
rendered. For example, if cell at (0, 3) would be rendered at (3, 0)
instead. The bug was difficult to find because of the nature of the 2D
array, where Cartesian coordinates are essentially backwards (row,
column) is essentially equivalent to (y, x).